Using weak_ptr to assoc. Model in Face
This commit solves the memory leak bug caused by circular references of
Model objects. The reason of this was the use of shared_ptr as a
reference to the associated Model object inside Face objects. When a
Face instance was put into their Model, the Model contained shared_ptrs
to itself so RAII destruction was impossible. With weak_ptrs inside of
Faces this is no longer a problem.